Rename package matryoshka_tester -> bci_tester
This fixes https://github.com/SUSE/BCI-tests/issues/36
