+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include bci_tester/data/containers.json
