+++ pyproject.toml
@@ -7,7 +7,6 @@ requires = [
 build-backend = 'setuptools.build_meta'
 
 [tool.setuptools_scm]
-#write_to = 'matryoshka_tester/version.py'
 
 [tool.pytest.ini_options]
 markers = [
