+++ setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
-name = matryoshka-tester
+name = bci-tester
 
 [options]
-packages = matryoshka_tester
+packages = bci_tester
 include_package_data = True
 
 [options.entry_points]
 console_scripts =
-    m8s-list-containers = matryoshka_tester.cmds:list_containers
-    m8a-fetch-all-containers = matryoshka_tester.cmds:fetch_all_containers
-    m8a-fetch-language-containers = matryoshka_tester.cmds:fetch_language_containers
-    m8a-trigger-tests-from-mqtt = matryoshka_tester.trigger_test:trigger_from_mqtt
+    bci-list-containers = bci_tester.cmds:list_containers
+    bci-fetch-all-containers = bci_tester.cmds:fetch_all_containers
+    bci-fetch-language-containers = bci_tester.cmds:fetch_language_containers
+    bci-trigger-tests-from-mqtt = bci_tester.trigger_test:trigger_from_mqtt
