+++ tox.ini
@@ -24,7 +24,7 @@ passenv =
     USER
     XDG_CONFIG_HOME
 commands =
-    m8a-fetch-language-containers {envname}
+    bci-fetch-language-containers {envname}
     # -n auto tells pytest to use as many processors as detected
     # to run all the tests from the module test_{envname} in parallel
     pytest -n auto -m "not serial" test_{envname}.py --junitxml={toxinidir}/junit_parallel.xml []
@@ -36,21 +36,21 @@ commands =
 
 [testenv:fetch-all]
 commands =
-    m8a-fetch-all-containers
+    bci-fetch-all-containers
 
 [testenv:list-all]
 commands =
-    m8s-list-containers
+    bci-list-containers
 
 [testenv:format]
 deps =
     black
 commands =
-    black . matryoshka_tester []
+    black . bci_tester []
 
 [testenv:trigger-tests]
 commands =
-    m8a-trigger-tests-from-mqtt
+    bci-trigger-tests-from-mqtt
 passenv = GITHUB_TOKEN
 
 [testenv:venv]
